requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
comm==0.2.2
cryptography==43.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1
django-cors-headers==4.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
executing==2.1.0
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.3.2
scipy==1.14.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
